+++ packages/specifications/src/candidature/corrigerCandidature.feature
@@ -12,12 +12,20 @@ Fonctionnalité: Corriger une candidature
         Et le porteur n'a pas été prévenu que son attestation a été modifiée
 
     Scénario: Corriger une candidature notifiée en regénérant l'attestation
-        Quand le DGEC validateur notifie la candidature comme lauréate
+        Etant donné la candidature lauréate notifiée "Du boulodrome de Marseille #2"
         Quand un administrateur corrige la candidature avec :
             | nom candidat               | abcd |
             | doit régénérer attestation | oui  |
         Alors la candidature devrait être consultable
         Et le porteur a été prévenu que son attestation a été modifiée
+
+    Scénario: Corriger une candidature notifiée sans régénérer l'attestation
+        Etant donné la candidature lauréate notifiée "Du boulodrome de Marseille #2"
+        Quand un administrateur corrige la candidature avec :
+            | nom candidat               | abcd |
+            | doit régénérer attestation | non  |
+        Alors la candidature devrait être consultable
+        Et le porteur n'a pas été prévenu que son attestation a été modifiée
 
     Scénario: Impossible de régénérer l'attestation d'une candidature non notifiée
         Quand un administrateur corrige la candidature avec :
