+++ packages/specifications/src/période/stepDefinitions/période.then.ts
@@ -1,6 +1,6 @@
 import { Then as Alors } from '@cucumber/cucumber';
 import waitForExpect from 'wait-for-expect';
-import { assert } from 'chai';
+import { assert, expect } from 'chai';
 import { mediator } from 'mediateur';
 
 import { Option } from '@potentiel-libraries/monads';
@@ -9,8 +9,10 @@ import { Période } from '@potentiel-domain/periode';
 import { Lauréat } from '@potentiel-domain/laureat';
 import { Éliminé } from '@potentiel-domain/elimine';
 import { Candidature } from '@potentiel-domain/candidature';
+import { ConsulterDocumentProjetQuery } from '@potentiel-domain/document';
 
 import { PotentielWorld } from '../../potentiel.world';
+import { convertReadableStreamToString } from '../../helpers/convertReadableToString';
 
 Alors(
   `la période devrait être notifiée avec les lauréats et les éliminés`,
@@ -28,6 +30,37 @@ Alors(
       await vérifierLauréats.call(this, this.périodeWorld.identifiantPériode);
       await vérifierÉliminés.call(this, this.périodeWorld.identifiantPériode);
     });
+  },
+);
+
+Alors(
+  'les attestations de désignation des candidatures de la période notifiée devraient être consultables',
+  async function (this: PotentielWorld) {
+    const { identifiantPériode } = this.périodeWorld;
+    const candidatures = await mediator.send<Candidature.ListerCandidaturesQuery>({
+      type: 'Candidature.Query.ListerCandidatures',
+      data: {
+        appelOffre: identifiantPériode.appelOffre,
+        période: identifiantPériode.période,
+      },
+    });
+
+    for (const candidature of candidatures.items) {
+      expect(candidature.attestation, "La candidature n'a pas d'attestation").not.to.be.undefined;
+      await waitForExpect(async () => {
+        if (candidature.attestation) {
+          const { content, format } = await mediator.send<ConsulterDocumentProjetQuery>({
+            type: 'Document.Query.ConsulterDocumentProjet',
+            data: {
+              documentKey: candidature.attestation.formatter(),
+            },
+          });
+
+          expect(await convertReadableStreamToString(content)).to.have.length.gt(1);
+          format.should.be.equal('application/pdf');
+        }
+      });
+    }
   },
 );
